rename packages to use entities on package name
