+++ jnosql-mapping/jnosql-mapping-graph/src/test/java/org/eclipse/jnosql/mapping/graph/AbstractGraphConverterTest.java
@@ -19,12 +19,12 @@ import org.apache.tinkerpop.gremlin.structure.Graph;
 import org.apache.tinkerpop.gremlin.structure.Property;
 import org.apache.tinkerpop.gremlin.structure.T;
 import org.apache.tinkerpop.gremlin.structure.Vertex;
-import org.eclipse.jnosql.mapping.graph.model.Computer;
-import org.eclipse.jnosql.mapping.graph.model.Job;
-import org.eclipse.jnosql.mapping.graph.model.Money;
-import org.eclipse.jnosql.mapping.graph.model.Movie;
-import org.eclipse.jnosql.mapping.graph.model.Person;
-import org.eclipse.jnosql.mapping.graph.model.Worker;
+import org.eclipse.jnosql.mapping.graph.entities.Computer;
+import org.eclipse.jnosql.mapping.graph.entities.Job;
+import org.eclipse.jnosql.mapping.graph.entities.Money;
+import org.eclipse.jnosql.mapping.graph.entities.Movie;
+import org.eclipse.jnosql.mapping.graph.entities.Person;
+import org.eclipse.jnosql.mapping.graph.entities.Worker;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
