+++ jnosql-mapping/jnosql-mapping-graph/src/test/java/org/eclipse/jnosql/mapping/graph/AbstractGraphInheritanceConverterTest.java
@@ -18,12 +18,12 @@ import org.apache.tinkerpop.gremlin.structure.Edge;
 import org.apache.tinkerpop.gremlin.structure.Graph;
 import org.apache.tinkerpop.gremlin.structure.T;
 import org.apache.tinkerpop.gremlin.structure.Vertex;
-import org.eclipse.jnosql.mapping.graph.model.inheritance.EmailNotification;
-import org.eclipse.jnosql.mapping.graph.model.inheritance.LargeProject;
-import org.eclipse.jnosql.mapping.graph.model.inheritance.Project;
-import org.eclipse.jnosql.mapping.graph.model.inheritance.SmallProject;
-import org.eclipse.jnosql.mapping.graph.model.inheritance.SmsNotification;
-import org.eclipse.jnosql.mapping.graph.model.inheritance.SocialMediaNotification;
+import org.eclipse.jnosql.mapping.graph.entities.inheritance.EmailNotification;
+import org.eclipse.jnosql.mapping.graph.entities.inheritance.LargeProject;
+import org.eclipse.jnosql.mapping.graph.entities.inheritance.Project;
+import org.eclipse.jnosql.mapping.graph.entities.inheritance.SmallProject;
+import org.eclipse.jnosql.mapping.graph.entities.inheritance.SmsNotification;
+import org.eclipse.jnosql.mapping.graph.entities.inheritance.SocialMediaNotification;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
