+++ jnosql-mapping/jnosql-mapping-graph/src/test/java/org/eclipse/jnosql/mapping/graph/DefaultVertexTraversalTest.java
@@ -18,9 +18,9 @@ import jakarta.nosql.NonUniqueResultException;
 import jakarta.nosql.tck.test.CDIExtension;
 import org.apache.tinkerpop.gremlin.process.traversal.P;
 import org.apache.tinkerpop.gremlin.structure.T;
-import org.eclipse.jnosql.mapping.graph.model.Animal;
-import org.eclipse.jnosql.mapping.graph.model.Book;
-import org.eclipse.jnosql.mapping.graph.model.Person;
+import org.eclipse.jnosql.mapping.graph.entities.Animal;
+import org.eclipse.jnosql.mapping.graph.entities.Book;
+import org.eclipse.jnosql.mapping.graph.entities.Person;
 import org.junit.jupiter.api.Test;
 
 import java.util.ArrayList;
