+++ jnosql-mapping/jnosql-mapping-graph/src/test/java/org/eclipse/jnosql/mapping/graph/entities/inheritance/EmailNotification.java
@@ -12,49 +12,30 @@
  *
  *   Otavio Santana
  */
-package org.eclipse.jnosql.mapping.graph.model;
 
+package org.eclipse.jnosql.mapping.graph.entities.inheritance;
 
 import jakarta.nosql.mapping.Column;
-import jakarta.nosql.mapping.Convert;
+import jakarta.nosql.mapping.DiscriminatorValue;
 import jakarta.nosql.mapping.Entity;
 
 @Entity
-public class Worker {
+@DiscriminatorValue("Email")
+public class EmailNotification extends Notification {
 
     @Column
-    private String name;
+    private String email;
 
-    @Column
-    private Job job;
-
-    @Column("money")
-    @Convert(MoneyConverter.class)
-    private Money salary;
-
-    public String getName() {
-        return name;
+    public String getEmail() {
+        return email;
     }
 
-    public void setName(String name) {
-        this.name = name;
+    public void setEmail(String email) {
+        this.email = email;
     }
 
-    public Job getJob() {
-        return job;
+    @Override
+    public String send() {
+        return "Sending message to email sms to the email: " + email;
     }
-
-    public void setJob(Job job) {
-        this.job = job;
-    }
-
-    public Money getSalary() {
-        return salary;
-    }
-
-    public void setSalary(Money salary) {
-        this.salary = salary;
-    }
-
-
 }
