+++ jnosql-mapping/jnosql-mapping-graph/src/test/java/org/eclipse/jnosql/mapping/graph/query/DeleteQueryConverterTest.java
@@ -16,8 +16,7 @@ package org.eclipse.jnosql.mapping.graph.query;
 
 import jakarta.nosql.mapping.Converters;
 import jakarta.nosql.mapping.Repository;
-import org.assertj.core.api.Assertions;
-import org.eclipse.jnosql.mapping.graph.model.Person;
+import org.eclipse.jnosql.mapping.graph.entities.Person;
 import org.eclipse.jnosql.mapping.reflection.EntityMetadata;
 import org.eclipse.jnosql.mapping.reflection.EntitiesMetadata;
 import org.apache.tinkerpop.gremlin.structure.Edge;
